Move doc handling to module
